Adding Firebase
I added firebase according to a tutorial I was watching.
However, when I compared with what already existed it looks like there
was already something kind of analogous there.

To be honest this may be a regression in the code so I'm fine to revert this.
Or if we talk about the different methods we may see one is better than the other. Either way.